+++ src/routes/(editor)/editor/views/code.svelte
@@ -1,0 +1,7 @@
+<script>
+    import { globalComponentCollectionStore } from "../../../globals/globalstores.js";
+</script>
+
+<div class="container-fluid">
+    <pre>{JSON.stringify($globalComponentCollectionStore, null, 2)}</pre>
+</div>