+++ src/routes/(editor)/editor/views/variables.svelte
@@ -1,8 +1,8 @@
 <script>
     import { onMount, onDestroy } from "svelte";
     import { globalComponentCollectionStore, globalEditorPreferencesStore, globalVisibilityStore,
-        globalRightPanelContentStore, globalLeftPanelContentStore  } from "../globals/globalstores.js";
-    import { PanelDisplayStyles, MenuLocations, ScreenSizePx } from "../globals/globalconstants.js";
+        globalRightPanelContentStore, globalLeftPanelContentStore  } from "../../../globals/globalstores.js";
+    import { PanelDisplayStyles, MenuLocations, ScreenSizePx } from "../../../globals/globalconstants.js";
     import jQuery from "jquery";
 
     let globalComponentCollection = $globalComponentCollectionStore ?? [{}] ;
@@ -23,11 +23,18 @@
     let bsWidthXxl = "100%";
     let bsWidthContainer = "100%";
 
+    let bsRatioSm = 1;
+    let bsRatioMd = 1;
+    let bsRatioLg = 1;
+    let bsRatioXl = 1;
+    let bsRatioXxl = 1;
+  
+
     /**
      * Width of screen
      */
     let outerWidth = 0;
-
+    let cssEditorWidth = 1320;
     let editorPanelWidth = 0;
     let editorPanelHeight = 0;
 
@@ -41,8 +48,8 @@
      */
     function updateBootstrapBreakPoints(width){
 
-        var editorWidth = Math.min(parseInt(width.replace("px","")), parseInt(editorPanelWidth.toString().replace("px","")));
-
+        // cssEditorWidth = Math.min(parseInt(width.replace("px","")), parseInt(editorPanelWidth.toString().replace("px",""))).toString()+"px";
+        cssEditorWidth = parseInt(width.replace("px",""));
 
         var xxl = 1320;
         var xl = 1140;
@@ -57,89 +64,129 @@
             // bsWidthMd = "100%";
             // bsWidthSm = "100%";
             // bsWidthContainer = "100%";
-            var value = Math.max(xxl, editorWidth);
-                bsWidthXxl = value+"px";
-                bsWidthXl = value+"px";
-                bsWidthLg = value+"px";
-                bsWidthMd = value+"px";
-                bsWidthSm = value+"px";
-                bsWidthContainer = value+"px";
+            // var value = Math.max(xxl, editorWidth);
+            //     bsWidthXxl = value+"px";
+            //     bsWidthXl = value+"px";
+            //     bsWidthLg = value+"px";
+            //     bsWidthMd = value+"px";
+            //     bsWidthSm = value+"px";
+            //     bsWidthContainer = value+"px";
+
+            bsRatioSm = 1;
+            bsRatioMd = 1;
+            bsRatioLg = 1;
+            bsRatioXl = 1;
+            bsRatioXxl = 1;
+         
         }else{
 
-            if(editorWidth > 1400){
-                // var newWidth = Math.round(editorWidth / xxl);
-                // var value = Math.min(xxl, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
-                var value = Math.max(xxl, editorWidth);
-                bsWidthXxl = value+"px";
-                bsWidthXl = value+"px";
-                bsWidthLg = value+"px";
-                bsWidthMd = value+"px";
-                bsWidthSm = value+"px";
-                bsWidthContainer = value+"px";
-            }else if( editorWidth <= 1400 && editorWidth > 1200){
-                // var newWidth = Math.round(editorWidth / xl);
-                // var value = Math.min(xl, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
-                var value = Math.max(xl, editorWidth);
-                bsWidthXxl = "100%";
-                bsWidthXl = value+"px";
-                bsWidthLg = value+"px";
-                bsWidthMd = value+"px";
-                bsWidthSm = value+"px";
-                bsWidthContainer = value+"px";
-            }else if( editorWidth <= 1200 && editorWidth > 992){
-                // var newWidth = Math.round(editorWidth / lg);
-                // var value = Math.min(lg, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
-                var value = Math.max(lg, editorWidth);
-                bsWidthXxl = "100%";
-                bsWidthXl = "100%";
-                bsWidthLg = value+"px";
-                bsWidthMd = value+"px";
-                bsWidthSm = value+"px";
-                bsWidthContainer = value+"px";
-            }else if( editorWidth <= 992 && editorWidth > 768){
-                // var newWidth = Math.round(editorWidth / md);
-                // var value = Math.min(md, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
-                var value = Math.max(md, editorWidth);
-                bsWidthXxl = "100%";
-                bsWidthXl = "100%";
-                bsWidthLg = "100%";
-                bsWidthMd = value+"px";
-                bsWidthSm = value+"px";
-                bsWidthContainer = value+"px";
-            }else if( editorWidth <= 768 && editorWidth > 576){
-                // var newWidth = Math.round(editorWidth / sm);
-                // var value = Math.min(sm, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
-                // console.log("editorWidth <= 768 && editorWidth > 576 ||=> newWidth: "+newWidth.toString()+" | sm: "+sm.toString());
-                console.log("editorWidth <= 768 && editorWidth > 576 ||=> editorWidth: "+editorWidth.toString()+" | sm: "+sm.toString());
-                var value = Math.max(sm, editorWidth);
-                bsWidthXxl = "100%";
-                bsWidthXl = "100%";
-                bsWidthLg = "100%";
-                bsWidthMd = "100%";
-                bsWidthSm = value+"px";
-                bsWidthContainer = value+"px";
-            }else if( editorWidth <= 576){
-                console.log("editorWidth <= 576 ||=> editorWidth: "+editorWidth.toString()+" | sm: "+sm.toString());
-                var value = editorWidth;
-                bsWidthXxl = value+"px";
-                bsWidthXl = value+"px";
-                bsWidthLg = value+"px";
-                bsWidthMd = value+"px";
-                bsWidthSm = value+"px";
-                bsWidthContainer = value+"px";
-                // bsWidthXxl = "100%";
-                // bsWidthXl = "100%";
-                // bsWidthLg = "100%";
-                // bsWidthMd = "100%";
-                // bsWidthSm = "100%";
-                // bsWidthContainer = "100%";
-            }
+          
+            bsRatioSm = editorPanelWidth / sm;
+            bsRatioMd = editorPanelWidth / md;
+            bsRatioLg = editorPanelWidth / lg;
+            bsRatioXl = editorPanelWidth / xl;
+            bsRatioXxl = editorPanelWidth / xxl;
+           
+
+          
+
+            // if(editorWidth > 1400){
+            //     // var newWidth = Math.round(editorWidth / xxl);
+            //     // var value = Math.min(xxl, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
+
+            //     bsRatioSm = 1;
+            //     bsRatioMd = 1;
+            //     bsRatioLg = 1;
+            //     bsRatioXl = 1;
+            //     bsRatioXxl = 1;
+            //     bsRatioContainer = 1;
+
+
+            //     var value = Math.max(xxl, editorWidth);
+            //     bsWidthXxl = value+"px";
+            //     bsWidthXl = value+"px";
+            //     bsWidthLg = value+"px";
+            //     bsWidthMd = value+"px";
+            //     bsWidthSm = value+"px";
+            //     bsWidthContainer = value+"px";
+            // }else if( editorWidth <= 1400 && editorWidth > 1200){
+            //     // var newWidth = Math.round(editorWidth / xl);
+            //     // var value = Math.min(xl, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
+
+
+            //     bsRatioSm = 1;
+            //     bsRatioMd = 1;
+            //     bsRatioLg = 1;
+            //     bsRatioXl = 1;
+            //     bsRatioXxl = 1;
+            //     bsRatioContainer = 1;
+
+
+            //     var value = Math.max(xl, editorWidth);
+            //     bsWidthXxl = "100%";
+            //     bsWidthXl = value+"px";
+            //     bsWidthLg = value+"px";
+            //     bsWidthMd = value+"px";
+            //     bsWidthSm = value+"px";
+            //     bsWidthContainer = value+"px";
+            // }else if( editorWidth <= 1200 && editorWidth > 992){
+            //     // var newWidth = Math.round(editorWidth / lg);
+            //     // var value = Math.min(lg, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
+            //     var value = Math.max(lg, editorWidth);
+            //     bsWidthXxl = "100%";
+            //     bsWidthXl = "100%";
+            //     bsWidthLg = value+"px";
+            //     bsWidthMd = value+"px";
+            //     bsWidthSm = value+"px";
+            //     bsWidthContainer = value+"px";
+            // }else if( editorWidth <= 992 && editorWidth > 768){
+            //     // var newWidth = Math.round(editorWidth / md);
+            //     // var value = Math.min(md, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
+            //     var value = Math.max(md, editorWidth);
+            //     bsWidthXxl = "100%";
+            //     bsWidthXl = "100%";
+            //     bsWidthLg = "100%";
+            //     bsWidthMd = value+"px";
+            //     bsWidthSm = value+"px";
+            //     bsWidthContainer = value+"px";
+            // }else if( editorWidth <= 768 && editorWidth > 576){
+            //     // var newWidth = Math.round(editorWidth / sm);
+            //     // var value = Math.min(sm, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
+            //     // console.log("editorWidth <= 768 && editorWidth > 576 ||=> newWidth: "+newWidth.toString()+" | sm: "+sm.toString());
+            //     console.log("editorWidth <= 768 && editorWidth > 576 ||=> editorWidth: "+editorWidth.toString()+" | sm: "+sm.toString());
+            //     var value = Math.max(sm, editorWidth);
+            //     bsWidthXxl = "100%";
+            //     bsWidthXl = "100%";
+            //     bsWidthLg = "100%";
+            //     bsWidthMd = "100%";
+            //     bsWidthSm = value+"px";
+            //     bsWidthContainer = value+"px";
+            // }else if( editorWidth <= 576){
+            //     console.log("editorWidth <= 576 ||=> editorWidth: "+editorWidth.toString()+" | sm: "+sm.toString());
+            //     var value = editorWidth;
+            //     bsWidthXxl = value+"px";
+            //     bsWidthXl = value+"px";
+            //     bsWidthLg = value+"px";
+            //     bsWidthMd = value+"px";
+            //     bsWidthSm = value+"px";
+            //     bsWidthContainer = value+"px";
+            //     // bsWidthXxl = "100%";
+            //     // bsWidthXl = "100%";
+            //     // bsWidthLg = "100%";
+            //     // bsWidthMd = "100%";
+            //     // bsWidthSm = "100%";
+            //     // bsWidthContainer = "100%";
+            // }
+
+
+
+
+            
+
 
         }
 
-                //////////////////////////////////////////////////////////////////////////////////////
-                //////// TODO: CONTAINER OLCULERI TAMAM AMA YINEDE HER BIR COL ICIN ORAN HESAPLAMAM LAZIM. BOOTSTRAP BAŞTAN BAŞLANACAK
-                //////////////////////////////////////////////////////////////////////////////////////
+
 
     }
 
@@ -167,6 +214,18 @@
             jQuery(target).removeClass("hovered");
         });
 
+        // var list = jQuery("div").hasClass("col-sm-12");
+
+        // jQuery(".col-sm-12").map(function(){
+        //     jQuery(this).attr("width", editorPanelWidth);
+        // });
+
+        var els = document.getElementsByClassName("col-sm-12");
+
+        Array.prototype.forEach.call(els, function(el) {
+            el.setAttribute("width", editorPanelWidth);
+        });
+
     });
 
 
@@ -175,14 +234,23 @@
 <svelte:window bind:outerWidth={outerWidth}/>
 
 
+
 <div id="editorInnerPanel" bind:clientWidth={editorPanelWidth} bind:clientHeight={editorPanelHeight} class="clearfix" style='
     --editorWidth:{$globalEditorPreferencesStore.editorData.editorWidth};
+    --editorMouseOverColor:{$globalEditorPreferencesStore.editorTheme.editorMouseOverColor};
     --bsWidthSm:{bsWidthSm};
     --bsWidthMd:{bsWidthMd};
     --bsWidthLg:{bsWidthLg};
     --bsWidthXl:{bsWidthXl};
     --bsWidthXxl:{bsWidthXxl};
     --bsWidthContainer:{bsWidthContainer};
+    --bsRatioSm:{bsRatioSm};
+    --bsRatioMd:{bsRatioMd};
+    --bsRatioLg:{bsRatioLg};
+    --bsRatioXl:{bsRatioXl};
+    --bsRatioXxl:{bsRatioXxl};
+    --cssEditorWidth:{cssEditorWidth};
+    --editorPanelWidth:{editorPanelWidth};
 
 '
 >
@@ -190,19 +258,26 @@
 
 
     <div class="container-fluid d-inline">
-        <h1>Hello this is editor page</h1>
+        <h1>Hello this is variables page</h1>
         <button class="btn btn-info">Test Button</button>
     </div>
 
     <div class="my-3">
         --editorWidth:{$globalEditorPreferencesStore.editorData.editorWidth};
+        --testPanelWidth:{testPanelWidth};
+        <br/><br/>
         --bsWidthSm:{bsWidthSm};
         --bsWidthMd:{bsWidthMd};
         --bsWidthLg:{bsWidthLg};
         --bsWidthXl:{bsWidthXl};
         --bsWidthXxl:{bsWidthXxl};
         --bsWidthContainer:{bsWidthContainer};
-        --testPanelWidth:{testPanelWidth};
+        <br/><br/>
+        --bsRatioSm:{bsRatioSm};
+        --bsRatioMd:{bsRatioMd};
+        --bsRatioLg:{bsRatioLg};
+        --bsRatioXl:{bsRatioXl};
+        --bsRatioXxl:{bsRatioXxl};
     </div>
 
     <div class="container row">
@@ -352,112 +427,17 @@
         flex-wrap: wrap; */
     }
 
+    /* calc(var(--editorPanelWidth)* 25 / 100); */
+
+    .hovered{
+        border: 2px solid var(--editorMouseOverColor);
+    }
+
     /* .hovered{
         border: 2px solid aqua;
     } */
 
-    .container{
-    max-width: var(--bsWidthContainer) !important;
-    }
-    .container-sm{
-        max-width: var(--bsWidthSm) !important;
-    }
-    .container-md{
-        max-width: var(--bsWidthMd) !important;
-    }
-    .container-lg{
-        max-width: var(--bsWidthLg) !important;
-    }
-    .container-xl{
-        max-width: var(--bsWidthXl) !important;
-    }
-    .container-xxl{
-        max-width: var(--bsWidthXxl) !important;
-    }
-
-  .col-sm {
-    flex: 1 0 0%;
-  }
-  .row-cols-sm-auto > * {
-    flex: 0 0 auto;
-    width: auto;
-  }
-  .row-cols-sm-1 > * {
-    flex: 0 0 auto;
-    width: 100%;
-  }
-  .row-cols-sm-2 > * {
-    flex: 0 0 auto;
-    width: 50%;
-  }
-  .row-cols-sm-3 > * {
-    flex: 0 0 auto;
-    width: 33.3333333333%;
-  }
-  .row-cols-sm-4 > * {
-    flex: 0 0 auto;
-    width: 25%;
-  }
-  .row-cols-sm-5 > * {
-    flex: 0 0 auto;
-    width: 20%;
-  }
-  .row-cols-sm-6 > * {
-    flex: 0 0 auto;
-    width: 16.6666666667%;
-  }
-  .col-sm-auto {
-    flex: 0 0 auto;
-    width: auto;
-  }
-  .col-sm-1 {
-    flex: 0 0 auto;
-    width: 8.33333333%;
-  }
-  .col-sm-2 {
-    flex: 0 0 auto;
-    width: 16.66666667%;
-  }
-  .col-sm-3 {
-    flex: 0 0 auto;
-    width: 25%;
-  }
-  .col-sm-4 {
-    flex: 0 0 auto;
-    width: 33.33333333%;
-  }
-  .col-sm-5 {
-    flex: 0 0 auto;
-    width: 41.66666667%;
-  }
-  .col-sm-6 {
-    flex: 0 0 auto;
-    width: 50%;
-  }
-  .col-sm-7 {
-    flex: 0 0 auto;
-    width: 58.33333333%;
-  }
-  .col-sm-8 {
-    flex: 0 0 auto;
-    width: 66.66666667%;
-  }
-  .col-sm-9 {
-    flex: 0 0 auto;
-    width: 75%;
-  }
-  .col-sm-10 {
-    flex: 0 0 auto;
-    width: 83.33333333%;
-  }
-  .col-sm-11 {
-    flex: 0 0 auto;
-    width: 91.66666667%;
-  }
-  .col-sm-12 {
-    flex: 0 0 auto;
-    width: 100%;
-  }
+    
 
 
     /* sm: 540px,