+++ src/routes/+page.svelte
@@ -1,14 +1,24 @@
 <script>
-    import Editor from "../routes/editor/+page.svelte";
+    import Editor from "./(editor)/editor/+page.svelte";
 
     let frame;
 
+    // console.log(frame);
+
 </script>
 
-<Editor />
+<!-- <Editor /> -->
+<svelte:component this={Editor}/>
 
-<!-- <iframe bind:this={frame} src="/editor" title="Editor" class="w-100 h-100" ></iframe> -->
+<!-- <iframe id="frame" bind:this={frame} src="/editor" title="Editor" class="" ></iframe> -->
 
 
 
-<style></style>
+<style>
+
+    #frame{
+        width: 100%;
+        height: 100vh;
+    }
+
+</style>