+++ src/routes/menu/options/hover.svelte
@@ -1,7 +1,9 @@
 <script>
 
+
 </script>
 
 <div>
-    <slot></slot>
- </div>
+   <slot></slot>
+</div>
+