+++ src/routes/menu/widgets/hover.svelte
@@ -3,4 +3,6 @@
 </script>
 
 
-<slot></slot>
+<div>
+    <slot></slot>
+ </div>